requirements.txt
asn1crypto==1.5.1
pg8000==1.31.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
scramp==1.4.5
sendgrid==6.11.0
six==1.16.0
starkbank-ecdsa==2.2.0
